Changed cutscenes to being a first class (empty type) object, separate cutscene exporting
