+++ fast64_internal/oot/oot_level_classes.py
@@ -112,6 +112,16 @@ class OOTCSList:
 		self.fxStartFrame = 0
 		self.fxEndFrame = 0
 
+class OOTCutscene:
+	def __init__(self):
+		self.name = ""
+		self.csEndFrame = 100
+		self.csWriteTerminator = False
+		self.csTermIdx = 0
+		self.csTermStart = 99
+		self.csTermEnd = 100
+		self.csLists = []
+
 class OOTSceneTableEntry:
 	def __init__(self):
 		self.drawConfig = 0
@@ -152,6 +162,9 @@ class OOTScene:
 		self.cameraList = []
 
 		self.writeCutscene = False
+		self.csWriteType = "Embedded"
+		self.csWriteCustom = ""
+		self.csWriteObject = None
 		self.csEndFrame = 100
 		self.csWriteTerminator = False
 		self.csTermIdx = 0
